Add Prisma setup and routing for plan and container pages
