+++ .gitignore
@@ -1,0 +1,4 @@
+node_modules
+.env
+prisma/dev.db
+tsconfig.tsbuildinfo
