+++ package.json
@@ -6,19 +6,24 @@
   "scripts": {
     "dev": "vite",
     "build": "tsc -b && vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "server": "node server.js"
   },
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "three": "^0.160.0",
     "@react-three/drei": "^9.99.0",
-    "@react-three/fiber": "^8.15.11"
+    "@react-three/fiber": "^8.15.11",
+    "@prisma/client": "^5.15.0",
+    "express": "^4.18.2",
+    "react-router-dom": "^6.23.0"
   },
   "devDependencies": {
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
     "@vitejs/plugin-react": "^4.2.0",
+    "prisma": "^5.15.0",
     "typescript": "^5.4.0",
     "vite": "^5.2.0"
   }
