+++ prisma/schema.prisma
@@ -1,0 +1,19 @@
+generator client {
+  provider = "prisma-client-js"
+}
+
+datasource db {
+  provider = "sqlite"
+  url      = env("DATABASE_URL")
+}
+
+model HU {
+  id           String   @id
+  length_cm    Int
+  width_cm     Int
+  height_cm    Int
+  weight_kg    Int
+  stackable    Boolean
+  deliveryDate DateTime
+  place        String
+}
