+++ src/pages/OptiPlan.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+export default function OptiPlan() {
+  return (
+    <div className="layout">
+      <h1>OptiPlan</h1>
+      <p>
+        Transport Planning Optimisation, based on overall volume and delivery date and place to detect possible consolidation.
+      </p>
+    </div>
+  );
+}
